Move the html-cleaner to the test module
