+++ src/clean.rs
@@ -1,6 +1,4 @@
 use kuchiki::{NodeDataRef, NodeRef};
-use reqwest;
-use serde_json::{self, Value};
 
 use crate::html;
 
@@ -12,19 +10,4 @@ pub fn correct_h1(dom: &NodeRef) {
             html::node::replace_node(h1s[0].as_node(), h1s[1].as_node());
         }
     }
-}
-
-pub fn clean_html(code: String) -> Result<String, Box<dyn std::error::Error>> {
-    let client = reqwest::blocking::Client::new();
-
-    let res = client
-        .post("https://www.10bestdesign.com/dirtymarkup/api/html")
-        .form(&[("code", code), ("indent", "2".to_owned())])
-        .send()?
-        .text()?;
-
-    let val: Value = serde_json::from_str(&res)?;
-    let cleaned = val["clean"].as_str().unwrap();
-
-    Ok(cleaned.to_owned())
 }
